requirements.txt
flask
flask-cors
gunicorn
requests
python-dotenv
google-generativeai
flask-sqlalchemy